test: update Requirement test suite for record.py and table.py
