+++ tests/__mocks__.py
@@ -10,52 +10,6 @@
 # Standard Library Imports
 from datetime import date, timedelta
 
-# Third Party Imports
-from treelib import Tree
-
-MOCK_HAZARDS = {
-    1: {
-        "assembly_effect": "",
-        "assembly_hri": 20,
-        "assembly_hri_f": 4,
-        "assembly_mitigation": "",
-        "assembly_probability": "Level A - Frequent",
-        "assembly_probability_f": "Level A - Frequent",
-        "assembly_severity": "Medium",
-        "assembly_severity_f": "Medium",
-        "function_1": "",
-        "function_2": "",
-        "function_3": "",
-        "function_4": "",
-        "function_5": "",
-        "potential_cause": "",
-        "potential_hazard": "",
-        "remarks": "",
-        "result_1": 0.0,
-        "result_2": 0.0,
-        "result_3": 0.0,
-        "result_4": 0.0,
-        "result_5": 0.0,
-        "system_effect": "",
-        "system_hri": 20,
-        "system_hri_f": 20,
-        "system_mitigation": "",
-        "system_probability": "Level A - Frequent",
-        "system_probability_f": "Level A - Frequent",
-        "system_severity": "Medium",
-        "system_severity_f": "Medium",
-        "user_blob_1": "",
-        "user_blob_2": "",
-        "user_blob_3": "",
-        "user_float_1": 0.0,
-        "user_float_2": 0.0,
-        "user_float_3": 0.0,
-        "user_int_1": 0,
-        "user_int_2": 0,
-        "user_int_3": 0,
-    }
-}
-
 MOCK_STATUS = {
     1: {
         "cost_remaining": 284.98,
@@ -63,105 +17,6 @@ MOCK_STATUS = {
         "time_remaining": 125.0,
     },
     2: {"cost_remaining": 212.32, "date_status": date.today(), "time_remaining": 112.5},
-}
-
-MOCK_REQUIREMENTS = {
-    1: {
-        "derived": 0,
-        "description": "",
-        "figure_number": "",
-        "owner": 0,
-        "page_number": "",
-        "parent_id": 0,
-        "priority": 0,
-        "requirement_code": "REL.1",
-        "specification": "",
-        "requirement_type": 0,
-        "validated": 0,
-        "validated_date": date.today(),
-        "q_clarity_0": 0,
-        "q_clarity_1": 0,
-        "q_clarity_2": 0,
-        "q_clarity_3": 0,
-        "q_clarity_4": 0,
-        "q_clarity_5": 0,
-        "q_clarity_6": 0,
-        "q_clarity_7": 0,
-        "q_clarity_8": 0,
-        "q_complete_0": 0,
-        "q_complete_1": 0,
-        "q_complete_2": 0,
-        "q_complete_3": 0,
-        "q_complete_4": 0,
-        "q_complete_5": 0,
-        "q_complete_6": 0,
-        "q_complete_7": 0,
-        "q_complete_8": 0,
-        "q_complete_9": 0,
-        "q_consistent_0": 0,
-        "q_consistent_1": 0,
-        "q_consistent_2": 0,
-        "q_consistent_3": 0,
-        "q_consistent_4": 0,
-        "q_consistent_5": 0,
-        "q_consistent_6": 0,
-        "q_consistent_7": 0,
-        "q_consistent_8": 0,
-        "q_verifiable_0": 0,
-        "q_verifiable_1": 0,
-        "q_verifiable_2": 0,
-        "q_verifiable_3": 0,
-        "q_verifiable_4": 0,
-        "q_verifiable_5": 0,
-    },
-    2: {
-        "derived": 1,
-        "description": "Derived requirement #1 for base requirement #1.",
-        "figure_number": "",
-        "owner": 0,
-        "page_number": "",
-        "parent_id": 1,
-        "priority": 0,
-        "requirement_code": "REL.1.1",
-        "specification": "",
-        "requirement_type": 0,
-        "validated": 0,
-        "validated_date": date.today(),
-        "q_clarity_0": 0,
-        "q_clarity_1": 0,
-        "q_clarity_2": 0,
-        "q_clarity_3": 0,
-        "q_clarity_4": 0,
-        "q_clarity_5": 0,
-        "q_clarity_6": 0,
-        "q_clarity_7": 0,
-        "q_clarity_8": 0,
-        "q_complete_0": 0,
-        "q_complete_1": 0,
-        "q_complete_2": 0,
-        "q_complete_3": 0,
-        "q_complete_4": 0,
-        "q_complete_5": 0,
-        "q_complete_6": 0,
-        "q_complete_7": 0,
-        "q_complete_8": 0,
-        "q_complete_9": 0,
-        "q_consistent_0": 0,
-        "q_consistent_1": 0,
-        "q_consistent_2": 0,
-        "q_consistent_3": 0,
-        "q_consistent_4": 0,
-        "q_consistent_5": 0,
-        "q_consistent_6": 0,
-        "q_consistent_7": 0,
-        "q_consistent_8": 0,
-        "q_verifiable_0": 0,
-        "q_verifiable_1": 0,
-        "q_verifiable_2": 0,
-        "q_verifiable_3": 0,
-        "q_verifiable_4": 0,
-        "q_verifiable_5": 0,
-    },
 }
 
 MOCK_STAKEHOLDERS = {
@@ -288,9 +143,3 @@ MOCK_VALIDATIONS = {
         "time_variance": 0.0,
     },
 }
-
-MOCK_RQRMNT_TREE = Tree()
-MOCK_RQRMNT_TREE.create_node(tag="requirement", identifier=0, parent=None, data=None)
-MOCK_RQRMNT_TREE.create_node(tag="REL-0001", identifier=1, parent=0, data=None)
-MOCK_RQRMNT_TREE.create_node(tag="PRF-0002", identifier=2, parent=0, data=None)
-MOCK_RQRMNT_TREE.create_node(tag="FUN-0003", identifier=3, parent=0, data=None)
